backend/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
DateTime==5.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.1
flask-restx==1.3.2
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.1.4
mysql-connector-python==9.5.0
packaging==25.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
zope.interface==8.0.1
